+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # 更改日志 | CHANGE LOG
+
+- 从 `index.html` 文件中获得样式文件 `base.*.css` | Get style file `base.*.css` from `index.html` file.